requirements.txt
certifi==2019.6.16
chardet==3.0.4
Deprecated==1.2.6
dj-database-url==0.5.0
Django==2.2.3
django-crispy-forms==1.7.2
django-extensions==2.1.9
gunicorn==19.9.0
idna==2.8
newsapi-python==0.2.5
psycopg2==2.8.3
PyGithub==1.43.7
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
webencodings==0.5.1
whitenoise==4.1.2
wrapt==1.11.2
